pyproject.toml
[project]
name = "curso-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.6",
    "groq>=0.30.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
    "yfinance>=0.2.65",
]
